@@ -92,9 +92,9 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3 :: Only",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.14",
   "Programming Language :: Python",
   "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
   "Topic :: Internet :: WWW/HTTP",
@@ -144,16 +144,16 @@ dependencies = [
   "lxml>=5.4.0,<6.1",
   "mistletoe>=1.4.0,<1.6",
   "nh3>=0.2.20,<0.4",
-  "openpyxl>=3.1.0,<3.2",
+  "openpyxl>=3.1.5,<3.2",
   "packaging>=25,<26",
   "phply>=1.2.6,<1.3",
   "Pillow>=11.0.0,<13",
   "pycairo>=1.20.0",
   "Pygments>=2.19.0,<3.0",
-  "PyGObject>=3.40.1,<3.55.0",
+  "PyGObject>=3.54.5,<3.55.0",
   "pyicumessageformat>=1.0.0,<1.1",
   "pyparsing>=3.2.0,<3.3",
-  "python-dateutil>=2.8.2",
+  "python-dateutil>=2.9.0.post0",
   "qrcode>=8.2,<8.3",
   "rapidfuzz>=3.12.1,<3.15",
   "redis>=5.2.0,<8.0.0",
@@ -184,7 +184,7 @@ keywords = [
 license = "GPL-3.0-or-later"
 license-files = ["LICENSE"]
 name = "weblate"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 version = "5.15"
 
 [project.optional-dependencies]
@@ -208,8 +208,8 @@ gerrit = [
   "git-review>=2.4.0,<2.6.0"
 ]
 google = [
-  "google-cloud-storage>=2.18.2,<3.7",
-  "google-cloud-translate>=3.13.0,<4.0"
+  "google-cloud-storage>=3.4.0,<3.7",
+  "google-cloud-translate>=3.21.0,<4.0"
 ]
 ldap = [
   "django-auth-ldap>=4.6.0,<6.0.0"
@@ -224,7 +224,7 @@ openai = [
   "openai>=2.0,<3.0"
 ]
 postgres = [
-  "psycopg[binary]>=3.1.8,<4"
+  "psycopg[binary]>=3.2.11,<3.3"
 ]
 saml = [
   "python3-saml>=1.16.0"
@@ -505,10 +505,6 @@ filterwarnings = [
   "ignore:datetime.datetime.utcnow:DeprecationWarning:onelogin.saml2.utils",
   # Warning in the third-party module (openid)
   "ignore:defusedxml.cElementTree is deprecated, import from defusedxml.ElementTree instead.:DeprecationWarning:openid.oidutil",
-  # importlib warning on Python 3.11
-  "ignore:DynamicImporter:ImportWarning",
-  # TODO: Will be fixed in next PyGobject release.
-  "ignore:GioUnix was imported without specifying a version first:gi.PyGIWarning",
   # There is no better way to differentiate DATA_DIR for the xdist executed tests
   "ignore:Implicitly cleaning up <TemporaryDirectory.*/xdist-:ResourceWarning",
   # Happens via dateparser
@@ -577,7 +573,6 @@ ignore = [
   "E501",  # WONTFIX: we accept long strings (rest is formatted by black)
   "FBT",  # TODO: Boolean in function definition
   "FIX002",  # CONFIG: we use TODO
-  "PERF203",  # WONTFIX: This rule is only enforced for Python versions prior to 3.11
   "PLC0415",  # TODO: `import` should be at the top-level of a file
   "PLR2004",  # TODO: Magic value used in comparison, consider replacing 201 with a constant variable
   "PLR6301",  # TODO: Method could be a function, class method, or static method
